MCP_2/pyproject.toml
[project]
name = "mcp-2"
version = "1.0.0"
description = "MCP Web Scraping Agent - A Model Context Protocol compatible web scraping agent with AI capabilities"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "numpy>=2.3.4",
    "pandas>=2.3.3",
    "fastapi[standard]",
    "crawl4ai",
    "google-generativeai",
    "python-dotenv",
    "graphviz",
    "beautifulsoup4",
    "requests"
]
